Majority of the CSS is finished
The majority of the CSS is finished, just need to talk with the rest of
the team about what color the icons will be, need to place more images
and figure out where other items need to be placed. Also, some front
end coding needs to be reworked on the about and contact page(refer
back to wireframes) having to columns that float left and right. Should
be fairly quick.